LPJmL data class

A data container for LPJmL input and output. Container - because an LPJmLData object is an environment in which the data array as well as the meta data are stored after read_io(). The data array can be accessed via $data, the meta data via $meta. The enclosing environment is locked and cannot be altered by any other than the available modify methods to ensure its integrity and validity. Use base stats methods like print(), summary.LPJmLData() or plot.LPJmLData() to get insights and export methods like as_tibble() or as_raster() to export it into common working formats.

Active bindings

meta

LPJmLMetaData object to store corresponding meta data.

data

array containing the underlying data.

grid

Optional LPJmLData object containing the underlying grid.

Methods

Public methods

LPJmLData$add_grid()

LPJmLData$subset()

LPJmLData$transform()

LPJmLData$as_array()

LPJmLData$as_tibble()

LPJmLData$as_raster()

LPJmLData$as_terra()

LPJmLData$plot()

LPJmLData$length()

LPJmLData$dim()

LPJmLData$dimnames()

LPJmLData$summary()

LPJmLData$print()

LPJmLData$.__set_data__()

LPJmLData$.__set_grid__()

LPJmLData$new()

LPJmLData$clone()

Method add_grid()

Method to add a grid to an LPJmLData object. See also add_grid

Usage

LPJmLData$add_grid(...)

Arguments

...

See add_grid().

Method subset()

Method to use dimension names of LPJmLData$data array directly to subset each dimension to match the supplied vectors.

Usage

LPJmLData$subset(...)

Arguments

...

See subset.LPJmLData()

Method transform()

Method to transform inner LPJmLData$data array into another space or time format.

Usage

LPJmLData$transform(...)

Arguments

...

See transform().

Method as_array()

Method to coerce (convert) an LPJmLData object into an array.

Usage

LPJmLData$as_array(...)

Arguments

...

See as_array().

Method as_tibble()

Method to coerce (convert) an LPJmLData object into a tibble (modern data.frame).

Usage

LPJmLData$as_tibble(...)

Arguments

...

See as_tibble().

Method as_raster()

Method to coerce (convert) an LPJmLData object into a raster or brick object that can be used for any GIS-based raster operations.

Usage

LPJmLData$as_raster(...)

Arguments

...

See as_raster().

Method as_terra()

Method to coerce (convert) an LPJmLData object into a rast object that can be used for any GIS-based raster operations.

Usage

LPJmLData$as_terra(...)

Arguments

...

See as_terra().

Method plot()

Method to plot a time-series or raster map of an LPJmLData object.

Usage

LPJmLData$plot(...)

Arguments

...

See plot.LPJmLData().

Method length()

Method to get the length of the data array of an LPJmLData object.
See also length.

Usage

LPJmLData$length()

Method dim()

Method to get the dimensions of the data array of an LPJmLData object.
See also dim.

Usage

LPJmLData$dim()

Method dimnames()

Method to get the dimnames (list) of the data array of an LPJmLData object.

Usage

LPJmLData$dimnames(...)

Arguments

...

See dimnames.LPJmLData().

Method summary()

Method to get the summary of the data array of an LPJmLData object.

Usage

LPJmLData$summary(...)

Arguments

...

See [summary.LPJmLData()].

Method print()

Method to print the LPJmLData object.
See also print.

Usage

LPJmLData$print()

Method .__set_data__()

!Internal method only to be used for package development!

Usage

LPJmLData$.__set_data__(data)

Arguments

data

Data array.

Method .__set_grid__()

!Internal method only to be used for package development!

Usage

LPJmLData$.__set_grid__(grid)

Arguments

grid

An LPJmLData object holding grid coordinates.

Method new()

!Internal method only to be used for package development!

Usage

LPJmLData$new(data, meta_data = NULL)

Arguments

data

array with LPJmL data.

meta_data

An LPJmLMetaData object.

Method clone()

The objects of this class are cloneable with this method.

Usage

LPJmLData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.