lpjmlkit: Toolkit for Basic LPJmL Handling

A collection of basic functions to facilitate the work with the Dynamic Global Vegetation Model (DGVM) Lund-Potsdam-Jena managed Land (LPJmL) hosted at the Potsdam Institute for Climate Impact Research (PIK). It provides functions for performing LPJmL simulations, as well as reading, processing and writing model-related data such as inputs and outputs or configuration files.

See also

Author

Maintainer: Jannes Breier jannesbr@pik-potsdam.de (ORCID)

Authors:

Sebastian Ostberg ostberg@pik-potsdam.de (ORCID)

Stephen Björn Wirth wirth@pik-potsdam.de (ORCID)

Sara Minoli minoli@pik-potsdam.de (ORCID)

Fabian Stenzel stenzel@pik-potsdam.de (ORCID)

Christoph Müller cmueller@pik-potsdam.de (ORCID)